server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.0.0
upstash-redis>=1.0.0
openai>=1.55.0
anthropic>=0.40.0
